% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_feedback.R
\name{add_feedback}
\alias{add_feedback}
\title{Add feedback  lines to a gantt table}
\usage{
add_feedback(data_set, bar_type, type2extend = "delivery",
  start_date = "start_date", end_date = "end_date", duration = "duration",
  feed_time = 15)
}
\examples{
tasks <- c("Review literature", "Mung data", "Stats analysis", "Write Report")
example <- data.frame(
nome        = factor(tasks, levels = tasks),
start_date  = as.Date(c("2010-08-24", "2010-10-01", "2010-11-01", "2011-02-14")),
end_date    = as.Date(c("2010-10-31", "2010-12-14", "2011-02-28", "2011-04-30")),
is.critical = c(TRUE, FALSE, FALSE, TRUE)
)
ggantt(add_feedback(tasks))
}
