% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggantt.R
\name{ggantt}
\alias{ggantt}
\title{Construct gantt charts with ggplot2}
\usage{
ggantt(dataset, label = "nome", start = "start_date", end = "end_date",
  cor = "royalblue")
}
\examples{
tasks <- c("Review literature", "Mung data", "Stats analysis", "Write Report")
example <- data.frame(
nome        = factor(tasks, levels = tasks),
start_date  = as.Date(c("2010-08-24", "2010-10-01", "2010-11-01", "2011-02-14")),
end_date    = as.Date(c("2010-10-31", "2010-12-14", "2011-02-28", "2011-04-30")),
is.critical = c(TRUE, FALSE, FALSE, TRUE)
)
ggantt(tasks)
}
